Move to core and add comparison testing Ironically, the tests are untested yet
