+++ hypothesis-python/tests/nocover/test_database_agreement.py
@@ -8,13 +8,73 @@
 # v. 2.0. If a copy of the MPL was not distributed with this file, You can
 # obtain one at https://mozilla.org/MPL/2.0/.
 
+import datetime
 import os
 import shutil
 import tempfile
+from pathlib import Path
+from zipfile import Path as ZipPath
+from zipfile import ZipFile
+
+from hypothesis.database import (
+    DirectoryBasedExampleDatabase,
+    InMemoryExampleDatabase,
+    GithubArtifactDatabase,
+    _hash,
+)
+from hypothesis.stateful import Bundle, RuleBasedStateMachine, rule
 
 from hypothesis import strategies as st
-from hypothesis.database import DirectoryBasedExampleDatabase, InMemoryExampleDatabase
-from hypothesis.stateful import Bundle, RuleBasedStateMachine, rule
+
+
+class GithubArtifactMock:
+    def __init__(self, tempd: str):
+        self.tempd = tempd
+
+        # This is a hack to create a fake GA artifact
+        # Create a new empty zip file
+        artifact_directory = Path(self.tempd) / "github-artifacts"
+        artifact_directory.mkdir(parents=True, exist_ok=True)
+        # This file should trigger the cache mechanism
+        artifact_path = (
+            artifact_directory / f"{datetime.datetime.now().isoformat()}.zip"
+        )
+        # Create a zip file
+        with ZipFile(artifact_path, "w") as zip_file:
+            # Create a new zip file
+            zip_file.writestr(".mock", "")
+
+        self._ga_root = ZipPath(artifact_path)
+        self.db = GithubArtifactDatabase("mock", "mock", artifact_directory)
+        self.db._initialize_db()
+
+        # Check that we have the same artifact path
+        assert self.db._artifact == artifact_path
+
+    def _value_path(self, key: bytes, value: bytes):
+        return self.db._key_path(key) / _hash(value)
+
+    def mock_save(self, key: bytes, value: bytes):
+        self.db._key_path(key).mkdir(parents=True, exist_ok=True)
+        path: Path = self._value_path(key, value)
+        if not path.exists():
+            path.write_bytes(value)
+
+    def mock_move(self, src: bytes, dest: bytes, value: bytes):
+        if src == dest:
+            self.mock_save(src, value)
+            return
+        try:
+            self._value_path(src, value).rename(self._value_path(dest, value))
+        except OSError:
+            self.mock_delete(src, value)
+            self.mock_save(dest, value)
+
+    def mock_delete(self, key: bytes, value: bytes):
+        try:
+            self._value_path(key, value).unlink()
+        except OSError:
+            pass
 
 
 class DatabaseComparison(RuleBasedStateMachine):
@@ -23,10 +83,19 @@ class DatabaseComparison(RuleBasedStateMachine):
         self.tempd = tempfile.mkdtemp()
         exampledir = os.path.join(self.tempd, "examples")
 
-        self.dbs = [
+        self.ga = GithubArtifactMock(self.tempd)
+
+        self.rw_dbs = [
             DirectoryBasedExampleDatabase(exampledir),
             InMemoryExampleDatabase(),
             DirectoryBasedExampleDatabase(exampledir),
+        ]
+
+        self.r_dbs = [
+            DirectoryBasedExampleDatabase(exampledir),
+            InMemoryExampleDatabase(),
+            DirectoryBasedExampleDatabase(exampledir),
+            self.ga.db
         ]
 
     keys = Bundle("keys")
@@ -42,24 +111,27 @@ class DatabaseComparison(RuleBasedStateMachine):
 
     @rule(k=keys, v=values)
     def save(self, k, v):
-        for db in self.dbs:
+        for db in self.rw_dbs:
             db.save(k, v)
+        self.ga.mock_save(k, v)
 
     @rule(k=keys, v=values)
     def delete(self, k, v):
-        for db in self.dbs:
+        for db in self.rw_dbs:
             db.delete(k, v)
+        self.ga.mock_delete(k, v)
 
     @rule(k1=keys, k2=keys, v=values)
     def move(self, k1, k2, v):
-        for db in self.dbs:
+        for db in self.rw_dbs:
             db.move(k1, k2, v)
+        self.ga.mock_move(k1, k2, v)
 
     @rule(k=keys)
     def values_agree(self, k):
         last = None
         last_db = None
-        for db in self.dbs:
+        for db in self.r_dbs:
             keys = set(db.fetch(k))
             if last is not None:
                 assert last == keys, (last_db, db)
